Support unpooled database connections
Currently, Kill Bill supports the use of either c3p0 or HikariCP for
pooling database connections. In either case, database connections
are managed by software running in the same JVM that Kill Bill is
running in.

It would be nice to have the option of configuring Kill Bill so that
it doesn't try to pool database connections. This would make it
easier to use external database connection pools like PGBouncer.